Refs #22875: Fix windows warnings

+++ include/fastdds/rtps/builtin/data/ParticipantBuiltinTopicData.hpp
@@ -39,8 +39,8 @@ struct ParticipantBuiltinTopicData
             const VendorId_t vendor_id,
             const dds::DomainId_t domain_id,
             const RTPSParticipantAllocationAttributes& allocation)
-        : user_data(allocation.data_limits.max_user_data)
-        , properties(allocation.data_limits.max_properties)
+        : user_data(static_cast<uint16_t>(allocation.data_limits.max_user_data))
+        , properties(static_cast<uint32_t>(allocation.data_limits.max_properties))
         , metatraffic_locators(allocation.locators.max_unicast_locators, allocation.locators.max_multicast_locators)
         , default_locators(allocation.locators.max_unicast_locators, allocation.locators.max_multicast_locators)
         , vendor_id(vendor_id)
